requirements.txt
django
djangorestframework
datamol
pandas
